+++ src/Exceptions/UnauthorizedRole.php
@@ -12,7 +12,7 @@ class UnauthorizedRole extends UnauthorizedException
      * UnauthorizedPermission constructor.
      *
      * @param $statusCode
-     * @param string $message
+     * @param string|null $message
      * @param array $requiredRoles
      */
     public function __construct($statusCode, string $message = null, array $requiredRoles = [])