+++ tests/TestHelper.php
@@ -18,7 +18,7 @@ class TestHelper
      *
      * @return int
      */
-    public function testMiddleware($middleware, $parameter)
+    public function testMiddleware(string $middleware, object $parameter): int
     {
         try {
             return $middleware->handle(new Request(), function () {